removed timecontrol from snapshot again. moved some files to correct/new package.
