+++ config.json
@@ -12,7 +12,6 @@
     "30+0",
     "30+20",
     "120+30",
-    "120+0/40:60+0",
-    "c: 5+0/2:10+0/4:0+5"
+    "120+0"
   ]
 }
